ServidorServicioOCR/ServidorPython/server/ApiRestOCR+AI/requirements.txt
#pip install -r requirements.txt
#pip install --no-cache-dir -r requirements.txt
#.\venv\Scripts\actívate
#sphinx-build -b html source/ _build/html
#pyinstaller --onefile main.py


Flask
python-dotenv
pydantic
Pillow
openai==0.28.0
Werkzeug
pytz
flasgger
ratelimit
cachetools
requests
pytesseract
opencv-python
numpy
pdf2image
PyPDF2
flask-cors
waitress
